+++ src/Kohkimakimoto/Worker/Job/RuntimeJob.php
@@ -1,6 +1,5 @@
 <?php
-namespace Kohkimakimoto\Worker\Foundation;
-
+namespace Kohkimakimoto\Worker\Job;
 
 /**
  * RuntimJob
